Add files via upload
The VGG model is pretrained on the imagenet dataset and we are using the weights obtained after the training in order to save our training time. The state-of-the-art VGG model will be implemented here in just 6 lines of code which will give us (>97%) accuracy on our Dogs vs Cats dataset.The VGG16 model can be trained here on the sample dataset, while the original(which is quite big) can be downloaded from Kaggle.